requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
passlib
python-jose
python-multipart